+++ CivilWar/CivilWar Server/Types.h
@@ -1,15 +1,15 @@
 #pragma once
 
-typedef int Radius;
-typedef int Speed;
-typedef int Width;
-typedef int Height;
+typedef float Radius;
+typedef float Speed;
+typedef float Width;
+typedef float Height;
 
 typedef short Health;
 
 typedef struct point {
-	int x;
-	int y;
+	float x;
+	float y;
 } Point;
 
 enum DirectionX { LEFT, RIGHT, NONE_X };
